requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
python-whois>=0.8.0
shodan>=1.31.0
colorama>=0.4.6
tabulate>=0.9.0
validators>=0.22.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0